# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/clients/{id}/groups": {
            "get": {
                "description": "Позволяет получить все группы, в которых состоит клиент",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "GetClientGroups",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "client_id",
                        "name": "input",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/clienthandlers.GetGroupsRequestDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/clients/{id}/students": {
            "get": {
                "description": "Позволяет получить всех студентов клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "GetClientStudents",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "client_id",
                        "name": "input",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/clienthandlers.GetClientStudentsHandlerResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/students/createandgroup": {
            "post": {
                "description": "Транзакция на создания ученика с записью в student_groups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "StudentsAndGroups",
                "parameters": [
                    {
                        "description": "Данные для создания студента и привязке к группе",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/studenthandlers.%D0%A1reateStudentInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StudentResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/students/groupedstudents": {
            "post": {
                "description": "Позволяет получить сгруппированных учеников школы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "GroupedStudents",
                "parameters": [
                    {
                        "description": "Фильтры: group_id и school_id(из токена авто)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/studenthandlers.GroupedStudentsInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studenthandlers.GroupedStudentsOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "clienthandlers.GetClientStudentsHandlerResponseDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StudentResponseDTO"
                    }
                }
            }
        },
...
        "clienthandlers.GetGroupsRequestDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GroupResponseDTO"
                    }
                }
            }
        },
...
        "dto.GroupResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "entity_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.StudentCreateUpdateDTO": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "skill_level": {
                    "type": "string"
                }
            }
        },
...
        "dto.StudentResponseDTO": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "skill_level": {
                    "type": "string"
                }
            }
        },
...
        "studenthandlers.GroupedStudentsInputDTO": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                }
            }
        },
        "studenthandlers.GroupedStudentsOutputDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StudentCreateUpdateDTO"
                    }
                }
            }
        },
...
    }
}